+++ packages/eyes-common/test/unit/config/ImageMatchSettings.spec.js
@@ -9,14 +9,14 @@ describe('ImageMatchSettings', () => {
     it('empty instance', () => {
       const ims = new ImageMatchSettings();
       const expectedSerialization = '{"matchLevel":"Strict","ignoreCaret":true,"useDom":false,"enablePatterns":false,' +
-        '"ignoreDisplacement":false,"ignore":[],"layout":[],"strict":[],"content":[],"floating":[]}';
+        '"ignoreDisplacements":false,"ignore":[],"layout":[],"strict":[],"content":[],"floating":[]}';
       assert.strictEqual(JSON.stringify(ims), expectedSerialization, 'ImageMatchSettings serialization does not match!');
     });
 
     it('with modified exact and ignore caret', () => {
       const ims = new ImageMatchSettings({ matchLevel: MatchLevel.Content, exact: new ExactMatchSettings(), ignoreCaret: true });
       const expectedSerialization = '{"matchLevel":"Content","ignoreCaret":true,"useDom":false,"enablePatterns":false,' +
-        '"ignoreDisplacement":false,"exact":{"minDiffIntensity":0,"minDiffWidth":0,"minDiffHeight":0,"matchThreshold":0},' +
+        '"ignoreDisplacements":false,"exact":{"minDiffIntensity":0,"minDiffWidth":0,"minDiffHeight":0,"matchThreshold":0},' +
         '"ignore":[],"layout":[],"strict":[],"content":[],"floating":[]}';
       assert.strictEqual(JSON.stringify(ims), expectedSerialization, 'ImageMatchSettings serialization does not match!');
     });
@@ -25,7 +25,7 @@ describe('ImageMatchSettings', () => {
       const ims = new ImageMatchSettings();
       ims.setIgnoreRegions([new Region(10, 20, 30, 40)]);
       const expectedSerialization = '{"matchLevel":"Strict","ignoreCaret":true,"useDom":false,"enablePatterns":false,' +
-        '"ignoreDisplacement":false,"ignore":[{"left":10,"top":20,"width":30,"height":40,"coordinatesType":' +
+        '"ignoreDisplacements":false,"ignore":[{"left":10,"top":20,"width":30,"height":40,"coordinatesType":' +
         '"SCREENSHOT_AS_IS"}],"layout":[],"strict":[],"content":[],"floating":[]}';
       assert.strictEqual(JSON.stringify(ims), expectedSerialization, 'ImageMatchSettings serialization does not match!');
     });
@@ -33,7 +33,7 @@ describe('ImageMatchSettings', () => {
     it('with modified useDom and enablePatterns', () => {
       const ims = new ImageMatchSettings({ useDom: true, enablePatterns: false });
       const expectedSerialization = '{"matchLevel":"Strict","ignoreCaret":true,"useDom":true,"enablePatterns":false,' +
-        '"ignoreDisplacement":false,"ignore":[],"layout":[],"strict":[],"content":[],"floating":[]}';
+        '"ignoreDisplacements":false,"ignore":[],"layout":[],"strict":[],"content":[],"floating":[]}';
       assert.strictEqual(JSON.stringify(ims), expectedSerialization, 'ImageMatchSettings serialization does not match!');
     });
   });