+++ packages/visual-grid-client/src/sdk/configParams.js
@@ -30,5 +30,5 @@ module.exports = [
   'concurrency',
   'useDom',
   'enablePatterns',
-  'ignoreDisplacement',
+  'ignoreDisplacements',
 ];