+++ packages/visual-grid-client/src/sdk/wrapperUtils.js
@@ -46,7 +46,7 @@ function configureWrappers({
   agentId,
   useDom,
   enablePatterns,
-  ignoreDisplacement,
+  ignoreDisplacements,
   assumeEnvironment,
 }) {
   const batchInfo = new BatchInfo({id: batchId, name: batchName});
@@ -74,7 +74,7 @@ function configureWrappers({
     matchLevel !== undefined && wrapper.setMatchLevel(matchLevel);
     useDom !== undefined && wrapper.setUseDom(useDom);
     enablePatterns !== undefined && wrapper.setEnablePatterns(enablePatterns);
-    ignoreDisplacement !== undefined && wrapper.setIgnoreDisplacement(ignoreDisplacement);
+    ignoreDisplacements !== undefined && wrapper.setIgnoreDisplacements(ignoreDisplacements);
     parentBranchName !== undefined && wrapper.setParentBranchName(parentBranchName);
     branchName !== undefined && wrapper.setBranchName(branchName);
     proxy !== undefined && wrapper.setProxy(proxy);