+++ packages/visual-grid-client/test/it/openEyes.it.test.js
@@ -1438,7 +1438,7 @@ describe('openEyes', () => {
       cdt: [],
       useDom: true,
       enablePatterns: false,
-      ignoreDisplacement: false,
+      ignoreDisplacements: false,
     });
     checkWindow({
       url: '',
@@ -1447,10 +1447,10 @@ describe('openEyes', () => {
     const [results] = await close();
     expect(results[0].__checkSettings.getUseDom()).to.be.true;
     expect(results[0].__checkSettings.getEnablePatterns()).to.be.false;
-    expect(results[0].__checkSettings.getIgnoreDisplacement()).to.be.false;
+    expect(results[0].__checkSettings.getIgnoreDisplacements()).to.be.false;
     expect(results[1].__checkSettings.getUseDom()).to.be.undefined;
     expect(results[1].__checkSettings.getEnablePatterns()).to.be.undefined;
-    expect(results[1].__checkSettings.getIgnoreDisplacement()).to.be.undefined;
+    expect(results[1].__checkSettings.getIgnoreDisplacements()).to.be.undefined;
   });
 
   it('handles abort', async () => {